Migrate tests to their respective commands
RF all usage of deprecated commands and arguments.
